vector to array

+++ NES/dma.h
@@ -9,8 +9,8 @@ class Dma {
   public:
     void run(uint8_t data, Ppu *ppu, uint8_t *ram)
     {
-        size_t offset = data << 8;
-        for (size_t i = 0; i < 0x100; i++) {
+        uint64_t offset = data << 8;
+        for (uint64_t i = 0; i < 0x100; i++) {
             ppu->sprite_ram[i] = ram[offset];
             offset += 1;
         }
